pyproject.toml
[project]
authors = [
    {name = "Ivan Kuznetsov", email = "ikuznets93@gmail.com"},
]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "pytest-cov>=6.2.1",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.7",
]

[project.scripts]
gendiff = "gendiff.scripts.main:main"
